requirements.txt
langchain==0.3.25
langchain-community==0.3.24
langchain-chroma==0.2.4
langchain-openai==0.3.16
chromadb==1.0.9
openai==1.77.0
ragas==0.2.15
pandas==2.2.3
tabulate==0.9.0
pyyaml==6.0.2
python-dotenv==1.1.0
tiktoken==0.9.0
pypdf==5.4.0